renamed GetPublicationStatusService -> GetServicePublication
