+++ apps/io-services-cms-webapp/src/webservice/controllers/get-service-publication/index.ts
